refactored Program context
